align coupla var types for #83

+++ src/traveltimes.cpp
@@ -580,7 +580,7 @@ size_t iso::trace_back_first (
 size_t iso::trace_back_prev_index (
         const Iso & iso,
         const size_t & stn,
-        const size_t & departure_time,
+        const int & departure_time,
         const int & trip_id,
         const bool &minimise_transfers
         )
